typescript: Convert Comment to typescript
